requirements.txt
bumpversion>=0.5.0
dnspython>=1.16.0
Eve>=0.9.0
exrex>=0.10.5
Faker<3.0.0
flasgger>=0.9.0
Flask>=1.1.0
gunicorn>=19.9.0
Jinja2>=2.10
jsonschema>=2.6.0,<3.0.0
pandas>=0.25.0
py-bcrypt>=0.4
py-healthcheck>=1.9.0
pyjwt>=1.7.1
pytest>=5.2.0
pytest-cov>=2.8.0
pytest-flask>=0.15.0
pytest-mock>=1.11.2
pytest-pep8>=1.0.0
pytest-pylint>=0.14.0
pytest-runner>=5.2
simplejson>=3.16.0
Sphinx>=2.2.0
sphinx-rtd-theme>=0.4.0
Werkzeug==0.15.4
